Reorganize triggers to centralize in index.ts, add trigger wrapper
