+++ src/geocode.ts
@@ -1,17 +1,9 @@
-import { isStakeholderApproved } from './helper'
 import { HEADQUARTER_COLUMN_NUMBER, LOCATIONS_SERVED_COLUMN_NUMBER, LIST_DELIMETER, HEADQUARTER_COORDINATES_COLUMN_NUMBER, LOCATIONS_SERVED_COORDINATES_COLUMN_NUMBER } from './constants'
-import { syncApprovedSheet } from './rtdb'
 
 // Upon approval,
 // Geocodes the headquarters cell and list of communities served cell to coordinates
 // Also sends the updated approved stakeholders sheet to the RTDB
-export function onStakeholderApprovalGeocodeAndSync(e: GoogleAppsScript.Events.SheetsOnEdit) {
-  if (!isStakeholderApproved(e))
-    return
-
-  const sheet = e.source.getActiveSheet()
-  const row = e.range.getRow()
-  
+export function geocodeRow(sheet: GoogleAppsScript.Spreadsheet.Sheet, row: number) {
   // Grab addresses
   const headquarterAddress: string = sheet.getRange(row, HEADQUARTER_COLUMN_NUMBER).getValue()
   const locationsServedAddresses: string[] = sheet.getRange(row, LOCATIONS_SERVED_COLUMN_NUMBER).getValue().split(LIST_DELIMETER)
@@ -23,8 +15,6 @@ export function onStakeholderApprovalGeocodeAndSync(e: GoogleAppsScript.Events.S
   // Put them into coordinate cells in the format 'lat, long' and 'lat1, long1;lat2, long2'
   sheet.getRange(row, HEADQUARTER_COORDINATES_COLUMN_NUMBER).setValue(`${headquarterCoordinates[0]}, ${headquarterCoordinates[1]}`)
   sheet.getRange(row, LOCATIONS_SERVED_COORDINATES_COLUMN_NUMBER).setValue(locationsServedCoordinates.map(coordinates => `${coordinates[0]}, ${coordinates[1]}`).join(LIST_DELIMETER))
-
-  syncApprovedSheet()
 }
 
 // Converts an address list to a list of coodinates
